move settings that determine where an image lives to UrlConfig, and settings for the generated thumbnail to UrlGenerator
